Add openstack exporter
This patch adds a new base exported for openstack based services.
This patch also adds a cinderbackend exporter which allows us to
fetch stats on all of the cinder driver's get_volume_stats() reporting
to the cinder scheduler.  This will allow us to watch the
overcommit ratios in each of the deployed backends to find out when
they are getting close to the max overcommit ratio.  When the max
overcommit ratio is reached, then cinder won't allocate new volumes
on that backend.   We need to be alerted to this prior to this happening
so we can allocate new luns/datastores in the vcenter deployment.